// src/app/api/civitai.ts
export enum Period {
    AllTime = "AllTime",
    Year = "Year",
    Month = "Month",
    Week = "Week",
    Day = "Day"
}
// ...
export type CivitAIImageStats = {
    cryCount: number,
    laughCount: number,
    likeCount: number,
    dislikeCount: number,
    heartCount: number,
    commentCount: number
}

export type CivitAIImage = {
    id: number;
    url: string;
    hash: string,
    width: number,
    height: number,
    nsfw: boolean,
    nsfwLevel: string,
    createdAt: string,
    postId: number,
    stats: CivitAIImageStats;
    meta: {
      Size: string;
      seed: number;
      Model: string;
      steps: number,
      prompt:  string
      sampler: string,
      cfgScale: number,
      Version: string,
      "Clip skip": string,
      "Hires upscale": string,
      "Hires upscaler": string,
      negativePrompt: string;
      "Denoising strength": string;
    },
    username: string;
}